Separate Mask R-CNN configs into separate module.
